+++ doc/changelog.rst
@@ -3,6 +3,8 @@ Changelog
 
 This changelog references all backward incompatibilities as we introduce them:
 
+* **2012-06-13**: Added an extension for the Symfony Security component
+
 * **2012-05-31**: Made the ``BrowserKit``, ``CssSelector``, ``DomCrawler``,
   ``Finder`` and ``Process`` components optional dependencies. Projects that
   depend on them (e.g. through functional tests) should add those dependencies